refactor: PointService에서 코드의 복잡도가 높고 사용도가 많은 updatePoint를 UpdatePointService로 분리하여 사용 (#207)
